Make pos unsigned

+++ sw/peripherals-board/src/drivers/hwrev2/hwrev2_single_lidar_open_round.hpp
@@ -35,7 +35,7 @@ private:
     bool turning = false;
     int turnDir = 1;           // 1 for clockwise, -1 for counterclockwise. Not known at start
     
-    int pos = 90;    // variable to store the servo position  
+    uint8_t pos = 90;    // variable to store the servo position  
 
     // About IMU
     float yaw;
